Upgrade code to integrate with Totem upstream
So far, we were using a rather old version of Totem to develop this plugin.

Now, in order to integrate it with Totem, we need to upgrade the code so it can
be built with Totem upstream.

Totem has changed its way of handling plugins since the version we were using,
so this patch adds the required changes.